% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auta.R
\docType{data}
\name{auta}
\alias{auta}
\title{Auta Dataset}
\format{a data.frame with 2400 rows and 8 columns}
\usage{
data(auta)
}
\description{
A subset of an auta2012 dataset from the package PogromcyDanych.
}
\details{
The subsequent columns describe:
\itemize{
\item Marka Brand of the car
\item Model Model of the car
\item Cena Price
\item KM Horse power
\item Pojemnosc Volume of the engine
\item Przebieg Mileage
\item Paliwo Type of fuel
\item Produkcja Year of production
}
}
\examples{
library("Przewodnik")
summary(auta)
}
\keyword{datasets}

