% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daneO.R
\docType{data}
\name{daneO}
\alias{daneO}
\title{Crunching of Oncology Data}
\format{a data.frame with 97 rows and 9 columns}
\usage{
data(daneO)
}
\description{
The subsequent columns describe:
\itemize{
\item Wiek age
\item Rozmiar.guza cancer size
\item Wezly.chlonne lymph nodes
\item Nowotwor cancer
\item Receptory.estrogenowe estrogen receptors
\item Receptory.progesteronowe progesteron receptors
\item Niepowodzenia failures
\item Okres.bez.wznowy time of observation
\item VEGF vascular endothelial growth factor
}
}
\keyword{datasets}

