% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daneSoc.R
\docType{data}
\name{daneSoc}
\alias{daneSoc}
\title{Some sociodemographic data}
\format{a data.frame with 204 rows and 7 columns}
\usage{
data(daneSoc)
}
\description{
The subsequent columns describe:
\itemize{
\item wiek Age
\item wyksztalcenie Education
\item st.cywilny Martial status
\item plec Sex
\item praca Work status
\item cisnienie.skurczowe Blood pressure
\item cisnienie.rozkurczowe Blood pressure
}
}
\keyword{datasets}

