% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maratony.R
\docType{data}
\name{maratony}
\alias{maratony}
\title{Marathons in Warsaw / Poland}
\format{a data.frame with 70736 rows and 14 columns}
\usage{
data(maratony)
}
\description{
Data from Warsaw marathons 2013 - 2016. Based on https://www.orlenmarathon.pl/ database.
The subsequent columns describe:
\itemize{
\item nazwisko_imie First and Last name
\item nr.startowy Number
\item plec Gender
\item rok Year
\item biegi_uliczne Name of the marathon
\item czas_brutto Gross Time
\item czas_brutto_sec Gross Time in secs
\item czas_netto Net Time
\item czas_netto_sec Net Time in sec
\item grup_wiek Age group
\item kategoria Category
\item kraj_nazwa Country
\item miejsce_kat Place in Category
\item miejsce_open Place in Open
}
}
\details{
Prepared by Krzysztof Trajkowski.
}
\keyword{datasets}

