% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mieszkania.R
\docType{data}
\name{mieszkania}
\alias{mieszkania}
\title{Housing Prices Data}
\format{a data.frame with 200 rows and 5 columns}
\usage{
data(mieszkania)
}
\description{
The subsequent columns describe:
\itemize{
\item cena price
\item pokoi number of rooms
\item powierzchnia surface
\item dzielnica district
\item typ.budynku house type
}
}
\examples{
library("Przewodnik")
summary(mieszkania)
}
\keyword{datasets}

